Fix for pads not working

+++ src/us/rynet/akaifire/receivers/ControlChangeReceiver.java
@@ -13,15 +13,15 @@ public class ControlChangeReceiver extends MidiReceiver {
   }
 
   public void send(MidiMessage message, long timeStamp) {
-    byte status        = (byte)message.getStatus();
-    byte controlNumber = (byte)(message.getMessage()[1] & 0xFF);
-    byte value         = (byte)(message.getMessage()[2] & 0xFF);
+    int status        = (int)message.getStatus();
+    int controlNumber = (int)(message.getMessage()[1] & 0xFF);
+    int value         = (int)(message.getMessage()[2] & 0xFF);
 
     if (status == AkaiFire.CONTROL_START || status == AkaiFire.CONTROL_STOP) {
       // TODO: Should be ALL controls - rename setPressed to isChanging or something
       akaiFire.getControls().stream().filter(control -> (control instanceof Button) && control.getIndex() == controlNumber).forEach(control -> ((Button)control).setPressed(status == AkaiFire.CONTROL_START));
     } else if (status == AkaiFire.KNOB_CHANGED) {
-      akaiFire.getControls().stream().filter(control -> (control instanceof Knob) && control.getIndex() == controlNumber).forEach(control -> ((Knob)control).midiChange(value));
+      akaiFire.getControls().stream().filter(control -> (control instanceof Knob) && control.getIndex() == controlNumber).forEach(control -> ((Knob)control).midiChange((byte)value));
     }
   }
